Remove session expiry calculations and token refresh code
There was a problem with my session refresh code that was wiping tokens,
but further to that, my session calculation code was assuming UTC when
it turns out that PHPIPAM does not issue session tokens in UTC, but in
the machine's local time (and my test machines just happened to be UTC).

This makes the session code not very useful, so I'm removing it for now.
Instead, client.Request will handle 403 errors due to token expiry and
attempt to re-login once, if that fails, Request will die.

Will re-add this code if necessary and if timezone becomes available in
time stamping.