controllers/subnets: Move mask to JSONIntString, remove State
Moved Mask to JSONIntString, as this field seems to contain no data when
the subnet is actually a folder (as per my tests on default data).

Also, removed the State field from subnet.Subnet. As per
phpipam/phpipam#756, this field is still currently in development and
hence cannot be considered stable. It appears that as of 1.2, the
controller accepts "state", but gives back "tag", and rejects "tag" on
PATCH/etc as an unknown field.