client/subnets: add support for custom fields, acceptance test SQL
This patch adds custom field support for subnets. It also contains core
functionality to read custom field schemas, and set custom fields on a
general scale. This means that all that really remains is to implement
the functionality across the other two controllers that support custom
fields (addresses and vlans).

Also included in this patch is a SQL script that will perform
post-installation configuration on a fresh PHPIPAM 1.2 install to
prepare it for acceptance testing. As acceptance testing is becoming
more complex, with the addition of custom fields, this is necessary to
ensure that a configuration is set up thew way we expect it to look when
performing tests.