controllers/subnets: Remove custom field from mock subnet JSON
This was needed if we were going to unit test the subnet custom field
read and update functions, but as these are are complex and make two
calls (by passing up to the appropriate functions in client.Client), we
can't test these functions properly with httptest. We lean on
integration testing instead, with unit coverage happening in the
upstream functions in the client package.