Custom fields cannot be required while using this SDK
Unfortunately, due to the design of PHPIPAM (explained in the README
insertions), required custom fields cannot be used by this SDK. People
will need to use optional fields with sane defaults.

Update the SQL script for acceptance testing as well so that the custom
fields are not provisioned required. We may need to fix a few acceptance
tests.