controllers/addresses: Add custom field support
This adds custom field support for the addresses controller. Adjustments
needed to be made to the test fixture while we were doing this, and we
have had to make notes that this SDK does not currently work with
required custom fields, so past tests needed to be adjusted as well.
Hopefully we don't have to do much extra work for the vlans controller
now that this is out of the way.